test(liveslots-tools): extract exo utils
